backend/api/requirements.txt
fastapi==0.105.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
requests-mock==1.11.0
uvicorn==0.24.0.post1
uvloop==0.19.0
azure-core==1.32.0
azure-cosmos==4.9.0
openai==1.61.1